Reorder extension authors
I absolutely do not mean to under-credit the original extension creator, but
they haven't had any part in the development of the extension for eight months.
I just think that if someone were to reach out to one of the authors via email,
they would probably start with the first one, which, with L.Longheval being the
first one, probably wouldn't yield a response.