Find available JDTLS versions on downloads page (#34)
Find available JDTLS versions on its downloads page as opposed to
checking GitHub tags, since they apparently aren't always synced with
what's actually available to download.

The implementation of this is... a little bit terrible in my opinion,
but it works. Hopefully this'll be rewritten to be cleaner and more
optimized later.

Resolves #33.