make runnable tasks shell agnostic (#93)
force /bin/sh in case the users default shell is not a POSIX-shell (e.g.
fish, nushell). This is a quick fix, it would be better to provide
proper scripts somehow, and/or make them user-definable.

Changed the scripts to use `./mvnw` instead of assuming a global
`mvn` command in case of maven projects.

Note the included tasks still do not work in windows