Remove python launch script dependency  (#97)
- For the JDTLS that the extension downloads and manages, we now also
include the launch script in the extension instead of using the one
provided by jdtls (which needs python).
- For users with `jdtls(.bat)` on their $PATH, it's still used as
before.

Apart from making it easy to use the extension for people who can't or
don't want to have python, this also allows us to manage the cache in
the future (clearing caches is sadly a common interaction with java
language servers)