Use our own fork of java-debug (#98)
Closes #67

Zed and java-debug appear to disagree in their interpretations of the
DAP-spec. Since we have no control over the release cycle of java-debug
and indeed whether they accept the change to begin with, and the Zed
team declines to adopt a more lenient stance in their interpretation of
the spec, we must run our own fork of java-debug.

I've created a fork with the required fixes and a release for us to use
here: https://github.com/playdohface/java-debug

This PR will make the extension download the forked release once if not
installed yet, and then use it indefinitely. When the underlying issues
are resolved either from Zeds side
(https://github.com/zed-industries/zed/issues/37080) or java-debug
merges our changes and makes a new release
(https://github.com/microsoft/java-debug/pull/609) we can switch back
our code to use the latest official release for java-debug.