Allow Java files to be run without requiring a build tool (#134)
- do not require a package declaration for runnables matcher
- compile and run with javac/java from $PATH if no build tool is found

This PR aims to improve the experience of using Java in Zed, mainly for
newcomers to both the language and/or the editor. With this, we make the
experience closer to what you find in VSCode or IntelliJ IDEA when it
comes to running a Java program. It's true that, at the moment, the
extension offers convenient ways of running a Java program and running
tests through extension-provided tasks that can either be selected from
the tasks menu or by clicking the runnable buttons in the gutter.

However, in order to use these convenient functionalities, the user
needs to be using a build tool in their project, be it Maven or Gradle.
Otherwise, they may find that they need to create custom tasks to obtain
a similar experience or just use the JDK CLI. This is a rather a
compromising assumption in my opinion, and it creates friction for
people that are trying Zed or Java or both for the first time and might
not be familiarized with build tools, Zed tasks or using the language
CLI to run their program. Or are just a bit lazy (like me) and want a
quick way to try something trivial without having to think about any of
those things.

The scope of these changes are limited to running a Java program, since
this is the most common task for the kind of scenarios that I think
benefit the main target audience of these changes. This means the test
runnables remain unchanged. Only the `main` runnables are touched in
this case.

So, to summarize the changes:

1. Previously, a file with a main method was only runnable if it
declared a package. Now, the `package_declaration` in the `main`
runnable queries was made optional, allowing us to generate runnables
for any number of files with a main method, even when a package is not
declared.
2. Previously, you could only run a Java program if you were using a
build tool (either Maven or Gradle). Now, we improve and modified the
task to run a Java program so that it compiles the `.java` files with
`javac`, put them in the `bin/` classpath, and finally run them with
`java` as a fallback for the scenarios when no build tools are present.

Finally, I tested this in projects using Maven, Gradle and no build tool
at all. Everything worked as expected. Feel free to check it out
yourselves though. Happy to hear any feeback!